academic-agent/requirements.txt
langgraph==0.0.30
langchain==0.0.300
openai==1.3.7
fastapi==0.95.2
uvicorn==0.22.0
requests==2.31.0
python-dotenv==1.0.0
faiss-cpu==1.7.4
sentence-transformers==2.2.2
scikit-learn==1.2.2
numpy==1.23.5
streamlit==1.25.0
PyYAML==6.0
altair==5.0.1
pandas==1.5.3
rapidfuzz==2.0.0

